3.Software/MIDI_Tranceiver/requirements.txt
mido
python-rtmidi
pyserial